#73: Convert CID to SQL statements   * used ':memory:' instead of a locale database file   * renamed variables   * improved the method 'as_sql_create_table' of the module interface so that the given sql dialect is in the list of the supported ones   * moved the method 'as_sql_create_table' to the sql module   * moved all sql tests from the fields and interface to the new module 'test_sql'
