+++ cutplace/interface.py
@@ -26,7 +26,7 @@ import inspect
 import io
 import logging
 import os.path
-import sqlite3 as lite
+import sqlite3
 
 import six
 
@@ -35,6 +35,7 @@ from cutplace import fields
 from cutplace import errors
 from cutplace import checks
 from cutplace import rowio
+from cutplace import sql
 from cutplace import _compat
 from cutplace import _tools
 from cutplace._compat import python_2_unicode_compatible
@@ -507,45 +508,6 @@ class Cid(object):
         """
         assert check_name is not None
         return self._check_name_to_check_map[check_name]
-
-    def as_sql_create_table(self, dialect='ansi'):
-        file_name = os.path.basename(self._cid_path)
-        table_name = file_name.split('.')
-
-        create_table = "create table " + table_name[0] + " (\n"
-        constraints = ""
-
-        # get column definitions and constraints for all fields
-        for field in self._field_formats:
-            column_def, constraint = field.as_sql(dialect)
-            create_table += column_def + ",\n"
-
-            if len(constraint) > 0:
-                constraints += constraint + ",\n"
-
-        constraints = constraints.rsplit(',', 1)[0]
-
-        create_table += constraints
-
-        create_table += "\n);"
-
-        con = None
-
-        try:
-            con = lite.connect("database.db")
-            cur = con.cursor()
-            cur.execute(create_table)
-
-        except lite.Error as err:
-            return err
-
-        finally:
-            if con:
-                cur = con.cursor()
-                cur.execute("drop table "+table_name[0]+" ;")
-                con.close()
-
-        return create_table
 
 
 def create_cid_from_string(cid_text):
